+++ tests/Numeric/DecimalSpec.hs
@@ -170,6 +170,8 @@ specBouned ::
      , Bounded a
      , NFData a
      , Round RoundHalfUp a
+     , Round RoundHalfDown a
+     , Round RoundHalfEven a
      , Round RoundDown a
      , Round RoundToZero a
      )
@@ -231,6 +233,8 @@ specRounding ::
      , Typeable p
      , Arbitrary p
      , Round RoundHalfUp p
+     , Round RoundHalfDown p
+     , Round RoundHalfEven p
      , Round RoundDown p
      , Round RoundToZero p
      )
@@ -239,6 +243,12 @@ specRounding = do
   prop (propNamePrefix . showsDecimalType @RoundHalfUp @(s + k) @p $ "") $
     prop_Rounding @RoundHalfUp @s @k @p
     (roundHalfUpTo (fromIntegral (natVal (Proxy :: Proxy s))))
+  prop (propNamePrefix . showsDecimalType @RoundHalfDown @(s + k) @p $ "") $
+    prop_Rounding @RoundHalfDown @s @k @p
+    (roundHalfDownTo (fromIntegral (natVal (Proxy :: Proxy s))))
+  prop (propNamePrefix . showsDecimalType @RoundHalfEven @(s + k) @p $ "") $
+    prop_Rounding @RoundHalfEven @s @k @p
+    (roundHalfEvenTo (fromIntegral (natVal (Proxy :: Proxy s))))
   prop (propNamePrefix . showsDecimalType @RoundToZero @(s + k) @p $ "") $
     prop_Rounding @RoundToZero @s @k @p
     (roundRoundToZeroTo (fromIntegral (natVal (Proxy :: Proxy s))))
@@ -386,9 +396,19 @@ assertExceptionIO isExc action =
 
 roundHalfUpTo :: Natural -> Rational -> Rational
 roundHalfUpTo to rational =
-  floor ((truncate (rational * ((s10 * 10) % 1) + 5) :: Integer) % 10) % s10
+  floor ((rational * ((s10 * 10) % 1) + 5) * (1 % 10)) % s10
   where
     s10 = 10 ^ to :: Integer
+
+roundHalfDownTo :: Natural -> Rational -> Rational
+roundHalfDownTo to rational =
+  ceiling ((rational * ((s10 * 10) % 1) - 5) * (1 % 10)) % s10
+  where
+    s10 = 10 ^ to :: Integer
+
+roundHalfEvenTo :: Natural -> Rational -> Rational
+roundHalfEvenTo to rational =
+  fromInteger (round $ rational * (10 ^ to)) / 10 ^ to
 
 roundFloorTo :: Natural -> Rational -> Rational
 roundFloorTo to rational = (floor (rational * (s10 % 1)) :: Integer) % s10
@@ -413,8 +433,3 @@ _roundCommercial to rational
              then q + 1
              else q) %
           s10
-
--- Once round half even is implemented, this can be used for testing
-_roundBankerTo :: Integer -> Rational -> Rational
-_roundBankerTo to rational =
-  fromInteger (round $ rational * (10 ^ to)) / 10 ^ to